Moved Cached interface back to Flow\ETL namespace
