+++ src/core/etl/src/Flow/ETL/Cache/Implementation/FilesystemCache.php
@@ -4,10 +4,10 @@ declare(strict_types=1);
 
 namespace Flow\ETL\Cache\Implementation;
 
-use Flow\ETL\Cache\{Cache, CacheIndex};
+use Flow\ETL\Cache\{CacheIndex};
 use Flow\ETL\Exception\KeyNotInCacheException;
 use Flow\ETL\Hash\NativePHPHash;
-use Flow\ETL\{Row, Rows};
+use Flow\ETL\{Cache, Row, Rows};
 use Flow\Filesystem\{Filesystem, Path};
 use Flow\Serializer\Serializer;
