+++ src/core/etl/src/Flow/ETL/Cache/Implementation/InMemoryCache.php
@@ -4,9 +4,9 @@ declare(strict_types=1);
 
 namespace Flow\ETL\Cache\Implementation;
 
-use Flow\ETL\Cache\{Cache, CacheIndex};
+use Flow\ETL\Cache\{CacheIndex};
 use Flow\ETL\Exception\KeyNotInCacheException;
-use Flow\ETL\{Row, Rows};
+use Flow\ETL\{Cache, Row, Rows};
 
 final class InMemoryCache implements Cache
 {
