+++ src/core/etl/src/Flow/ETL/Cache/Implementation/PSRSimpleCache.php
@@ -4,9 +4,9 @@ declare(strict_types=1);
 
 namespace Flow\ETL\Cache\Implementation;
 
-use Flow\ETL\Cache\{Cache, CacheIndex};
+use Flow\ETL\Cache\{CacheIndex};
 use Flow\ETL\Exception\KeyNotInCacheException;
-use Flow\ETL\{Row, Rows};
+use Flow\ETL\{Cache, Row, Rows};
 use Flow\Serializer\{NativePHPSerializer, Serializer};
 use Psr\SimpleCache\{CacheInterface, InvalidArgumentException};
