+++ src/core/etl/src/Flow/ETL/Config/Cache/CacheConfigBuilder.php
@@ -5,14 +5,14 @@ declare(strict_types=1);
 namespace Flow\ETL\Config\Cache;
 
 use function Flow\Filesystem\DSL\protocol;
-use Flow\ETL\Cache\{Cache, Implementation\FilesystemCache};
+use Flow\ETL\Cache\{Implementation\FilesystemCache};
 use Flow\ETL\Exception\{InvalidArgumentException, RuntimeException};
 use Flow\Filesystem\{FilesystemTable, Path};
 use Flow\Serializer\Serializer;
 
 final class CacheConfigBuilder
 {
-    private ?Cache $cache = null;
+    private ?\Flow\ETL\Cache $cache = null;
 
     /**
      * @var int<1, max>
@@ -52,7 +52,7 @@ final class CacheConfigBuilder
         );
     }
 
-    public function cache(Cache $cache) : self
+    public function cache(\Flow\ETL\Cache $cache) : self
     {
         $this->cache = $cache;
